fix last minute critical bugs
- adding a scene
- deleting notes/cards with attached places/characters
- doc exporter
- confirmation before deletion of characters/notes/places/tags

+++ src/app/components/notes/noteView.js
@@ -29,7 +29,9 @@ class NoteView extends Component {
   }
 
   deleteNote = () => {
-    this.props.actions.deleteNote(this.props.note.id)
+    if (window.confirm(`Do you want to delete this note: '${this.props.note.title}'?`)) {
+      this.props.actions.deleteNote(this.props.note.id)
+    }
   }
 
   renderContent () {
